+++ src/main/java/br/com/loja/gabrielvinicius/services/ProdutoService.java
@@ -6,6 +6,8 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import br.com.loja.gabrielvinicius.excecoes.ExcecaoRecursoNaoEncontrado;
+import br.com.loja.gabrielvinicius.excecoes.ObjetoObrigatorioExcecaoNula;
 import br.com.loja.gabrielvinicius.models.Produto;
 import br.com.loja.gabrielvinicius.repositories.ProdutoRepository;
 
@@ -16,6 +18,7 @@ public class ProdutoService {
 	private ProdutoRepository produtoRepository;
 
 	public void salvarProduto(Produto produto) {
+		if (produto == null) throw new ObjetoObrigatorioExcecaoNula();
 		produtoRepository.save(produto);
 	}
 
@@ -24,17 +27,33 @@ public class ProdutoService {
 	}
 
 	public Optional<Produto> getProduto(int id) {
+		if (produtoRepository.findById(id) == null) { throw new ObjetoObrigatorioExcecaoNula();}
+		produtoRepository.findById(id)
+		.orElseThrow(() -> new ExcecaoRecursoNaoEncontrado("Nenhum registro encontrado para este ID!"));
 		return produtoRepository.findById(id);
+	}
+	
+	
+	public Produto updateProduto(Produto produto) {
+
+		if (produto == null) throw new ObjetoObrigatorioExcecaoNula();		
+		
+		var entidade = produtoRepository.findById(produto.getId())
+			.orElseThrow(() -> new ExcecaoRecursoNaoEncontrado("Nenhum registro encontrado para este ID!"));
+		produtoRepository.save(entidade);
+			
+		return entidade;
 	}
 
 	public void excluirProduto(Produto produto) {
+		if (produto == null) throw new ObjetoObrigatorioExcecaoNula();
 		produtoRepository.delete(produto);
 	}
 
 	public boolean verificarNome(String nome) {
 		List<Produto> produtos = listAll();
 		for (Produto produto : produtos) {
-			if (produto.getNome() == nome) {
+			if (produto.getNome().equals(nome)) {
 				return true;
 			}
 		}
@@ -58,17 +77,15 @@ public class ProdutoService {
 		return false;
 	}
 
-	public void diminuirEstoque(int quantidade, List<Produto> produtos) {
-		for (Produto produto : produtos) {
+	public void diminuirEstoque(int quantidade, Produto produto) {
 			Integer estoque = produto.getEstoque();
             estoque -= quantidade;
-        }
 	}
 
-	public Boolean validaProduto(String nome, int id, Produto produto) {
+	public Boolean validaProduto(String nome, int id, int estoque) {
 		List<Produto> produtos = listAll();
 		for (Produto p : produtos) {
-			if (nome == p.getNome() && id == p.getId() && produto.getEstoque() < 0)
+			if (p.getNome() == nome && p.getId() == id  && p.getEstoque() < 0)
 				return true;
 		}
 			return false;
