+++ src/main/java/br/com/loja/gabrielvinicius/services/VendaService.java
@@ -6,6 +6,8 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import br.com.loja.gabrielvinicius.excecoes.ExcecaoRecursoNaoEncontrado;
+import br.com.loja.gabrielvinicius.excecoes.ObjetoObrigatorioExcecaoNula;
 import br.com.loja.gabrielvinicius.models.Produto;
 import br.com.loja.gabrielvinicius.models.Venda;
 import br.com.loja.gabrielvinicius.repositories.ProdutoRepository;
@@ -21,8 +23,16 @@ public class VendaService {
 	private ProdutoRepository produtoRepository;
 
 	
+	
 	public void salvarVenda(Venda venda) {
-		if(verificaQuantidade(produtoRepository.findAll(), venda)) {
+		if (venda == null) { throw new ObjetoObrigatorioExcecaoNula();}
+		
+		Integer cliente =  venda.getCliente().getId();
+		Integer vendedor = venda.getVendedor().getId();
+		Boolean valida = venda.getDataVenda().after(venda.getDataGarantia());
+	    Boolean verificaQuantidade = verificaQuantidade(venda.getProdutos(), venda);
+	    
+		if (validaVenda(cliente, vendedor, valida, verificaQuantidade)) {
 		Double valorTotal = valorTotal(produtoRepository.findAll(), venda);
 		venda.setValorTotal(valorTotal);
 		vendaRepository.save(venda);
@@ -34,13 +44,26 @@ public class VendaService {
 	}
 
 	public Optional<Venda> getVenda(int id){
+		if (vendaRepository.findById(id) == null) { throw new ObjetoObrigatorioExcecaoNula();}
+		vendaRepository.findById(id)
+		.orElseThrow(() -> new ExcecaoRecursoNaoEncontrado("Nenhum registro encontrado para este ID!"));
 		return vendaRepository.findById(id);
 	}
 	
 	public void excluirVenda(Venda venda) {
+		if (venda == null) { throw new ObjetoObrigatorioExcecaoNula();}
 		vendaRepository.delete(venda);
 	}
-
+	
+	public Venda updateVenda(Venda venda) {
+		if (venda == null) throw new ObjetoObrigatorioExcecaoNula();		
+		
+		var entidade = vendaRepository.findById(venda.getId())
+			.orElseThrow(() -> new ExcecaoRecursoNaoEncontrado("Nenhum registro encontrado para este ID!"));
+		vendaRepository.save(entidade);	
+		return entidade;
+	}
+	
     public boolean verificarId(int id) {
     	List<Venda> vendaes = listAll();
         for (Venda venda : vendaes) {
@@ -54,15 +77,14 @@ public class VendaService {
 	public Double valorTotal(List<Produto> produtos, Venda venda) {	
 		Double total = null;
 		for (Produto produto : produtos) {
-			total =+ (produto.getPreco() * venda.getQuantidade());
+			total =+ (produto.getPreco() * produto.getQuantidade());
 		}
 		return total;
 	}
 	
 	public boolean verificaQuantidade(List<Produto> produtos, Venda venda) {
-		Integer quantidadeVenda = venda.getQuantidade();
 		for (Produto produto : produtos) {
-			if(produto.getEstoque() < quantidadeVenda) {
+			if(produto.getEstoque() < produto.getQuantidade()) {
 				 return true;
 			 }
 		}
